requirements.txt
fastapi[all]
torch
torchvision
requests
Pillow
redis
rq
python-multipart
matplotlib
python-magic
python-magic-bin